+++ mathics/builtin/numbers/calculus.py
@@ -26,7 +26,7 @@ from mathics.core.convert import sympy_symbol_prefix, SympyExpression, from_symp
 from mathics.core.rules import Pattern
 from mathics.core.numbers import dps
 from mathics.builtin.scoping import dynamic_scoping
-from mathics.builtin.numeric import _numeric_evaluation_with_prec
+from mathics.builtin.numeric import apply_N
 from mathics import Symbol
 
 import sympy
@@ -1280,7 +1280,7 @@ def find_root_newton(f, x0, x, opts, evaluation) -> (Number, bool):
         # TODO: use Precision goal...
         if x1 == x0:
             break
-        x0 = _numeric_evaluation_with_prec(x1, evaluation)
+        x0 = apply_N(x1, evaluation)
         # N required due to bug in sympy arithmetic
         count += 1
     else:
@@ -1376,7 +1376,7 @@ class FindRoot(Builtin):
     def apply(self, f, x, x0, evaluation, options):
         "FindRoot[f_, {x_, x0_}, OptionsPattern[]]"
         # First, determine x0 and x
-        x0 = _numeric_evaluation_with_prec(x0, evaluation)
+        x0 = apply_N(x0, evaluation)
         if not isinstance(x0, Number):
             evaluation.message("FindRoot", "snum", x0)
             return