+++ CHANGES.rst
@@ -6,7 +6,7 @@ Internals
 
 * now `Expression.is_numeric()` accepts an `Evaluation` object as a parameter,
   to use the definitions.
-* To numerify expressions, the function `_numeric_evaluation_with_prec` was introduced in the module `mathics.builtin.numeric` to avoid the idiom
+* To numerify expressions, the function `apply_N` was introduced in the module `mathics.builtin.numeric` to avoid the idiom
   `Expression("N", expr, prec).evaluate(evaluation)`. The idea is to avoid when it is possible to call the Pattern matching routines to obtain the numeric value of an expression.
 * A bug comming from a failure in the order in which `mathics.core.definitions` stores the rules was fixed.